Renamed files for `JoltHingeJointImpl3D`
