+++ src/servers/jolt_physics_server_3d.cpp
@@ -2,7 +2,7 @@
 
 #include "joints/jolt_cone_twist_joint_impl_3d.hpp"
 #include "joints/jolt_generic_6dof_joint_impl_3d.hpp"
-#include "joints/jolt_hinge_joint_3d.hpp"
+#include "joints/jolt_hinge_joint_impl_3d.hpp"
 #include "joints/jolt_joint_3d.hpp"
 #include "joints/jolt_pin_joint_3d.hpp"
 #include "joints/jolt_slider_joint_3d.hpp"
